Rearranges individual unit metrics to their own files
